requirements-py3.txt
atomicwrites==1.3.0
attrs==18.2.0
googleapis-common-protos==1.5.8
grpcio-tools==1.19.0
more-itertools>=5.0.0
pluggy==0.8.1
py==1.7.0
pytest==4.3.0
six==1.12.0
